refactor(auth): organize user and service boundaries
