+++ packages/control-plane/src/auth/authenticate.ts
@@ -3,287 +3,32 @@
  * typed `Principal` before any handler runs.
  *
  * A `sig1` service signature is verified against that service's own secret.
- * Browser requests additionally require a Better Auth session. Anything else
+ * User requests additionally require a Better Auth session. Anything else
  * is not a recognized credential.
  *
  * Sandbox tokens stay router-verified (they need the session id from the
  * path and a DO round-trip), so they are not dispatched here.
  */
 
-import {
-  ACTOR_HEADER,
-  SERVICE_HEADER,
-  SERVICE_SIGNATURE_HEADER,
-  TOKEN_VALIDITY_MS,
-  isServiceName,
-  parseServiceSignatureHeader,
-  readBodyCapped,
-  sha256Hex,
-  verifyServiceSignature,
-  type ServiceName,
-} from "@open-inspect/shared";
-
-import {
-  ASSERTION_RIGHTS,
-  isActorNamespace,
-  type ActorNamespace,
-  type AuthenticationContext,
-  type Principal,
-} from "./principal";
-import {
-  authenticateBrowserSession,
-  BrowserSessionIntegrityError,
-} from "./browser-session-authenticator";
-import { UserStore } from "../db/user-store";
+import { SERVICE_SIGNATURE_HEADER } from "@open-inspect/shared";
+import { authenticateSession, SessionIntegrityError } from "./user/session-authenticator";
+import { isAuthError, type AuthResult } from "./result";
+import { authenticateServiceRequest } from "./service/request-authenticator";
 import { createLogger } from "../logger";
 import type { RequestContext } from "../routes/shared";
 import type { Env } from "../types";
 
 const logger = createLogger("auth");
 
-export interface AuthError {
-  /** Response body message (also the log detail). Never carries token material. */
-  reason: string;
-  status: 401 | 413 | 500;
-  /**
-   * Which scheme was attempted and failed. A per-service attempt is terminal;
-   * "none" means no recognized credential was presented at all, and the
-   * router may still try sandbox auth on sandbox routes.
-   */
-  failedScheme: "per-service" | "browser-session" | "none";
-}
-
-/**
- * Hard cap on a service-signed request body. The signature covers the body
- * hash, so the body must be buffered and hashed before verification can
- * finish — this cap bounds what an unauthenticated sender can make the edge
- * buffer. The largest legitimate signed body is a session-attachment
- * multipart upload (see SESSION_ATTACHMENT_MAX_REQUEST_BYTES, ~10MB); sandbox
- * media uploads authenticate with sandbox tokens and never pass through here.
- */
-export const SERVICE_REQUEST_MAX_BODY_BYTES = 16 * 1024 * 1024;
-
-export type AuthResult =
-  | {
-      principal: Principal;
-      request: Request;
-      authentication?: AuthenticationContext;
-    }
-  | AuthError;
-
-export function isAuthError(result: AuthResult): result is AuthError {
-  return !("principal" in result);
-}
+export { isAuthError, type AuthError, type AuthResult } from "./result";
+export { SERVICE_REQUEST_MAX_BODY_BYTES } from "./service/request-authenticator";
 
 export interface AuthenticationRequirement {
   /**
    * Whether a verified service:web request is acting as the service itself or
-   * as a browser user through a Better Auth session.
+   * as a user through a Better Auth session.
    */
-  readonly webService?: "service" | "browser";
-}
-
-/** The per-service verification keys the CP holds. */
-export interface ServiceKeyEnv {
-  SERVICE_AUTH_SECRET_WEB?: string;
-  SERVICE_AUTH_SECRET_SLACK_BOT?: string;
-  SERVICE_AUTH_SECRET_GITHUB_BOT?: string;
-  SERVICE_AUTH_SECRET_LINEAR_BOT?: string;
-  SERVICE_AUTH_SECRET_MODAL?: string;
-}
-
-/** The verification key the CP holds for each service (also the signing key for CP→bot callbacks). */
-export function serviceAuthSecret(env: ServiceKeyEnv, service: ServiceName): string | undefined {
-  switch (service) {
-    case "web":
-      return env.SERVICE_AUTH_SECRET_WEB;
-    case "slack-bot":
-      return env.SERVICE_AUTH_SECRET_SLACK_BOT;
-    case "github-bot":
-      return env.SERVICE_AUTH_SECRET_GITHUB_BOT;
-    case "linear-bot":
-      return env.SERVICE_AUTH_SECRET_LINEAR_BOT;
-    case "modal":
-      return env.SERVICE_AUTH_SECRET_MODAL;
-  }
-}
-
-/** Parse `<namespace>:<id>` into a typed actor reference; null when malformed. */
-function parseActor(actor: string): { provider: ActorNamespace; providerUserId: string } | null {
-  const separator = actor.indexOf(":");
-  if (separator <= 0) return null;
-  const namespace = actor.slice(0, separator);
-  const providerUserId = actor.slice(separator + 1);
-  if (providerUserId === "" || !isActorNamespace(namespace)) return null;
-  return { provider: namespace, providerUserId };
-}
-
-/**
- * Best-effort nonce-reuse detection (log-only for now; a future change may
- * reject). In-isolate only — a replay against a different isolate is not
- * observed. Entries expire with the signature validity window.
- */
-const seenNonces = new Map<string, number>();
-const SEEN_NONCE_LIMIT = 5000;
-
-function recordNonce(service: ServiceName, nonce: string, ctx: RequestContext): void {
-  const now = Date.now();
-  const key = `${service}:${nonce}`;
-  const expiresAt = seenNonces.get(key);
-  if (expiresAt !== undefined && expiresAt > now) {
-    logger.warn("Service auth nonce reused", {
-      event: "auth.nonce_reuse",
-      service,
-      request_id: ctx.request_id,
-      trace_id: ctx.trace_id,
-    });
-    return;
-  }
-  if (seenNonces.size >= SEEN_NONCE_LIMIT) {
-    for (const [candidate, expiry] of seenNonces) {
-      if (expiry <= now) seenNonces.delete(candidate);
-    }
-    // Still over cap: shed the oldest entries (Map iteration is insertion
-    // order, and entries are inserted with monotonically increasing expiry).
-    // A flood must degrade detection gradually, never erase all memory.
-    let excess = seenNonces.size - SEEN_NONCE_LIMIT + 1;
-    for (const candidate of seenNonces.keys()) {
-      if (excess-- <= 0) break;
-      seenNonces.delete(candidate);
-    }
-  }
-  seenNonces.set(key, now + TOKEN_VALIDITY_MS);
-}
-
-async function authenticateServiceCredential(
-  request: Request,
-  env: Env,
-  ctx: RequestContext,
-  signatureHeader: string
-): Promise<AuthResult> {
-  const serviceHeader = request.headers.get(SERVICE_HEADER) ?? "";
-  if (!isServiceName(serviceHeader)) {
-    logger.warn("Service auth failed: unknown service", {
-      event: "auth.service_failed",
-      failure: "unknown_service",
-      service: serviceHeader,
-      request_id: ctx.request_id,
-      trace_id: ctx.trace_id,
-    });
-    return { reason: "Unauthorized", status: 401, failedScheme: "per-service" };
-  }
-  const service = serviceHeader;
-
-  const secret = serviceAuthSecret(env, service);
-  if (!secret) {
-    logger.error("Service auth secret not configured - rejecting request", {
-      event: "auth.misconfigured",
-      service,
-      request_id: ctx.request_id,
-      trace_id: ctx.trace_id,
-    });
-    return {
-      reason: "Service authentication not configured",
-      status: 500,
-      failedScheme: "per-service",
-    };
-  }
-
-  // Reject everything rejectable before paying for the body: a malformed or
-  // stale header must not cost a body buffer + hash.
-  const parsedSignature = parseServiceSignatureHeader(signatureHeader);
-  if (!parsedSignature.ok) {
-    logger.warn("Service auth failed: signature rejected", {
-      event: "auth.service_failed",
-      failure: parsedSignature.reason,
-      service,
-      request_id: ctx.request_id,
-      trace_id: ctx.trace_id,
-    });
-    return { reason: "Unauthorized", status: 401, failedScheme: "per-service" };
-  }
-
-  let bodyBuffer: Uint8Array | null = null;
-  if (request.body !== null) {
-    bodyBuffer = await readBodyCapped(request.body, SERVICE_REQUEST_MAX_BODY_BYTES);
-    if (bodyBuffer === null) {
-      logger.warn("Service auth failed: body over size cap", {
-        event: "auth.service_failed",
-        failure: "body_too_large",
-        service,
-        request_id: ctx.request_id,
-        trace_id: ctx.trace_id,
-      });
-      return { reason: "Request body too large", status: 413, failedScheme: "per-service" };
-    }
-  }
-  const bodySha256Hex = await sha256Hex(bodyBuffer ?? "");
-  const actor = request.headers.get(ACTOR_HEADER) ?? "";
-
-  const verification = await verifyServiceSignature({
-    signatureHeader,
-    service,
-    secret,
-    method: request.method,
-    url: request.url,
-    bodySha256Hex,
-    actor,
-  });
-  if (!verification.ok) {
-    logger.warn("Service auth failed: signature rejected", {
-      event: "auth.service_failed",
-      failure: verification.reason,
-      service,
-      request_id: ctx.request_id,
-      trace_id: ctx.trace_id,
-    });
-    return { reason: "Unauthorized", status: 401, failedScheme: "per-service" };
-  }
-
-  recordNonce(service, verification.nonce, ctx);
-
-  let resolvedActor = null;
-  if (actor !== "") {
-    const parsed = parseActor(actor);
-    if (!parsed || ASSERTION_RIGHTS[service] !== parsed.provider) {
-      logger.warn("Actor assertion denied", {
-        event: "auth.assertion_denied",
-        service,
-        actor,
-        request_id: ctx.request_id,
-        trace_id: ctx.trace_id,
-      });
-      return { reason: "Unauthorized", status: 401, failedScheme: "per-service" };
-    }
-    const identity = await new UserStore(ctx.db).getIdentity(
-      parsed.provider,
-      parsed.providerUserId
-    );
-    resolvedActor = {
-      provider: parsed.provider,
-      providerUserId: parsed.providerUserId,
-      canonicalUserId: identity?.userId ?? null,
-      participantUserId: actor,
-    };
-  }
-
-  // The body was consumed to hash it; hand the handler a request that can
-  // still be read. Bodyless requests pass through untouched. Built from
-  // parts deliberately: the `new Request(request, { body })` copy-constructor
-  // throws in workerd once the source request's body has been disturbed.
-  const handlerRequest =
-    bodyBuffer === null
-      ? request
-      : new Request(request.url, {
-          method: request.method,
-          headers: request.headers,
-          body: bodyBuffer,
-        });
-
-  return {
-    principal: { kind: "service", service, actor: resolvedActor },
-    request: handlerRequest,
-  };
+  readonly webService?: "service" | "user";
 }
 
 export async function authenticate(
@@ -294,33 +39,30 @@ export async function authenticate(
 ): Promise<AuthResult> {
   const signatureHeader = request.headers.get(SERVICE_SIGNATURE_HEADER);
   if (signatureHeader !== null) {
-    const channel = await authenticateServiceCredential(request, env, ctx, signatureHeader);
+    const channel = await authenticateServiceRequest(request, env, ctx, signatureHeader);
     if (
       isAuthError(channel) ||
       channel.principal.kind !== "service" ||
       channel.principal.service !== "web" ||
-      requirement.webService !== "browser"
+      requirement.webService !== "user"
     ) {
       return channel;
     }
-    if (!ctx.getBrowserAuth) {
-      logger.error("Browser authentication runtime unavailable", {
+    if (!ctx.getUserAuth) {
+      logger.error("User authentication runtime unavailable", {
         event: "auth.browser.misconfigured",
         request_id: ctx.request_id,
         trace_id: ctx.trace_id,
       });
       return {
-        reason: "Browser authentication is not configured",
+        reason: "User authentication is not configured",
         status: 500,
         failedScheme: "browser-session",
       };
     }
     try {
-      const browser = await authenticateBrowserSession(
-        ctx.getBrowserAuth(),
-        channel.request.headers
-      );
-      if (!browser) {
+      const userSession = await authenticateSession(ctx.getUserAuth().api, channel.request.headers);
+      if (!userSession) {
         return {
           reason: "Unauthorized",
           status: 401,
@@ -328,20 +70,20 @@ export async function authenticate(
         };
       }
       return {
-        principal: { kind: "user", userId: browser.userId },
-        authentication: browser.authentication,
+        principal: { kind: "user", userId: userSession.userId },
+        authentication: userSession.authentication,
         request: channel.request,
       };
     } catch (cause) {
-      logger.error("Browser session validation failed", {
+      logger.error("User session validation failed", {
         event: "auth.browser.failed",
-        failure: cause instanceof BrowserSessionIntegrityError ? "integrity" : "runtime",
+        failure: cause instanceof SessionIntegrityError ? "integrity" : "runtime",
         error: cause,
         request_id: ctx.request_id,
         trace_id: ctx.trace_id,
       });
       return {
-        reason: "Browser authentication failed",
+        reason: "User authentication failed",
         status: 500,
         failedScheme: "browser-session",
       };
