--- packages/control-plane/src/auth/index.ts
@@ -1,7 +1,0 @@
-/**
- * Auth module exports.
- */
-
-export { encryptToken, decryptToken, generateEncryptionKey, generateId } from "./crypto";
-
-export { isGitHubAppConfigured, getGitHubAppConfig, type GitHubAppConfig } from "./github-app";
