+++ packages/control-plane/src/auth/user/session-authenticator.ts
@@ -1,5 +1,5 @@
 /**
- * Adapts a Better Auth browser session into the control plane's authentication model.
+ * Adapts a Better Auth session into the control plane's authentication model.
  *
  * `authenticate()` calls this module only after validating the `service:web`
  * sig1 channel. This adapter delegates opaque session-cookie verification and
@@ -10,12 +10,11 @@
  *
  * Protected resource requests use a non-refreshing session read so validation
  * does not extend session expiry or write D1 as a side effect. Browser-facing
- * Better Auth session endpoints remain responsible for session refresh.
+ * Better Auth endpoints remain responsible for session refresh.
  */
 
 import { z } from "zod";
-import type { AuthenticationContext } from "./principal";
-import type { BrowserAuthRuntime } from "./browser-auth-runtime";
+import type { AuthenticationContext } from "../principal";
 
 const sessionSchema = z.object({
   session: z.object({
@@ -27,24 +26,31 @@ const sessionSchema = z.object({
   }),
 });
 
-export interface AuthenticatedBrowserUser {
+export interface AuthenticatedUserSession {
   readonly userId: string;
   readonly authentication: AuthenticationContext;
 }
 
-export class BrowserSessionIntegrityError extends Error {
+export interface SessionReader {
+  getSession(input: {
+    readonly headers: Headers;
+    readonly query: { readonly disableRefresh: true };
+  }): Promise<unknown>;
+}
+
+export class SessionIntegrityError extends Error {
   constructor(message: string) {
     super(message);
-    this.name = "BrowserSessionIntegrityError";
+    this.name = "SessionIntegrityError";
   }
 }
 
-export async function authenticateBrowserSession(
-  auth: BrowserAuthRuntime,
+export async function authenticateSession(
+  sessionReader: SessionReader,
   headers: Headers
-): Promise<AuthenticatedBrowserUser | null> {
+): Promise<AuthenticatedUserSession | null> {
   // Resource authentication is a read-only hot path, not a session-lifecycle endpoint.
-  const candidate = await auth.api.getSession({
+  const candidate = await sessionReader.getSession({
     headers,
     query: { disableRefresh: true },
   });
@@ -52,11 +58,11 @@ export async function authenticateBrowserSession(
 
   const parsedSession = sessionSchema.safeParse(candidate);
   if (!parsedSession.success) {
-    throw new BrowserSessionIntegrityError("Better Auth returned a malformed session");
+    throw new SessionIntegrityError("Better Auth returned a malformed session");
   }
   const { session, user } = parsedSession.data;
   if (session.userId !== user.id) {
-    throw new BrowserSessionIntegrityError("Better Auth returned a cross-user session");
+    throw new SessionIntegrityError("Better Auth returned a cross-user session");
   }
 
   return {
