+++ packages/control-plane/src/auth/user/sign-in-provider.ts
@@ -1,4 +1,4 @@
-/** Browser sign-in providers with executable authentication adapters. */
+/** Sign-in providers with executable authentication adapters. */
 export const SIGN_IN_PROVIDERS = ["github", "google"] as const;
 
 export type SignInProvider = (typeof SIGN_IN_PROVIDERS)[number];
