+++ packages/control-plane/src/routes/shared.ts
@@ -9,7 +9,7 @@ import type { RequestMetrics } from "../db/instrumented-d1";
 import type { SqlDatabase } from "../db/sql-database";
 import type { Env } from "../types";
 import type { Logger } from "../logger";
-import type { BrowserAuthRuntime } from "../auth/browser-auth-runtime";
+import type { BetterAuthRuntime } from "../auth/user/runtime";
 import {
   createSourceControlProviderFromEnv,
   SourceControlProviderError,
@@ -31,8 +31,8 @@ export type RequestContext = CorrelationContext & {
   db: SqlDatabase;
   /** Worker ExecutionContext for waitUntil (background tasks). */
   executionCtx?: ExecutionContext;
-  /** Lazy runtime dependency used by browser-session authentication and credential access. */
-  getBrowserAuth?: () => BrowserAuthRuntime;
+  /** Lazy runtime dependency used by user-session authentication and credential access. */
+  getUserAuth?: () => BetterAuthRuntime;
   /**
    * The request's verified principal. Absent only on public routes and CORS
    * preflights — every authenticated request carries one.
